Allow OR-only clauses; add tests
This feature was (effectively) requested by Thomas Kahle: allow clauses
which only consist of ORs, without implications. If there are only ANDs,
you have to use multiple clauses, which was supported already.

This commit also adds a simple array of tests for the propositional
syntax and Seq machinery; and a new test for the OR-only clauses.