Support decomposition of conditioning sets
In a propositional formula the `cube` statement now allows an arbitrary
number of atoms and of subsets to be specified, e.g. `cube(ijkl|MNOP)`.
This will cause the axioms to be iterated over all assignments to the
atoms and all decompositions of the complement of the atoms to the sets.

Previously only one set was allowed, e.g. `cube(ijkl|M)`, which makes
the axioms a finite forbidden-minor specification. With this commit,
we allow more general axiomatic frameworks.