Explicitly `use feature 'current_sub'`
The function generated by the `propositional` keyword uses the __SUB__
feature without enabling it explicitly, thus relying on the user to have
enabled it beforehand (usually through some `use Modern::Perl`).

This creates trouble in the Eval::Closure sandbox that is used by
`CImake` in `CInet::Tools`, so better enable it explicitly.