Allow attributes in `propositional` and add Families
This commit allows attributes to be added to a `propositional` declaration
which are applied to the sub defined by the keyword. This can be used to
tag propositionals as :Export, for example.

This is used in the new CInet::Propositional::Families module which defines
commonly used families of CI relations which have a finite axiomatization:
Semigraphoids, Graphoids, Semigaussoids, Gaussoids and MarkovNetworks.